Moved most of tests to tests directory
